requirements_prod.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-dotenv==1.1.0
pydantic==2.11.4
dacite==1.9.2
websocket-client==1.8.0
certifi==2025.4.26
requests==2.32.3
typing-extensions==4.13.2 